move IPKernelApp from zmq.ipkernel to zmq.kernelapp
- merged IPKernelApp into KernelApp, they are no longer separate classes
- embed_kernel moved to its own file
- ipkernel now only contains the Kernel class
- associated imports updated
