+++ IPython/__init__.py
@@ -81,5 +81,5 @@ def embed_kernel(module=None, local_ns=None, **kwargs):
         local_ns = caller_locals
     
     # Only import .zmq when we really need it
-    from .zmq.ipkernel import embed_kernel as real_embed_kernel
+    from .zmq.embed import embed_kernel as real_embed_kernel
     real_embed_kernel(module=module, local_ns=local_ns, **kwargs)
